+++ docs/CHANGELOG.md
@@ -1,3 +1,17 @@
+# [2.26.0-beta.2](https://github.com/shdwmtr/millennium/compare/v2.26.0-beta.1...v2.26.0-beta.2) (2025-04-14)
+
+
+### Bug Fixes
+
+* Fix plugin updater sometimes extracting to wrong path, updating while plugin is running, and fixed backend memory management issues., ([7c47e88](https://github.com/shdwmtr/millennium/commit/7c47e887a948c2f83d51ba5108814f35e0ff62d0))
+
+# [2.26.0-beta.1](https://github.com/shdwmtr/millennium/compare/v2.25.1...v2.26.0-beta.1) (2025-04-13)
+
+
+### Features
+
+* Add plugin updater. ([5c28c7e](https://github.com/shdwmtr/millennium/commit/5c28c7ef96e576ace25a6be111535c6fa0768cdf))
+
 ## [2.25.1](https://github.com/shdwmtr/millennium/compare/v2.25.0...v2.25.1) (2025-04-12)